Multiplayer for Guessing game
Laid a foundation for multiplayer functionality for my Guessing Game. Worked with loop, switch, if-else, goto and struct, simultaneously to understand different concepts and make them work together to achieve my eventual goal. Which in this case, was to simply make the game multiplayer so that there could be something like a leaderboard at the end, which will induce a little more fun in one of the simplest game for any language, as a beginner.