Multiplayer Functionality
Finalized the initiation process of the game for multiplayers. The program now asks user of how many players will be playing? Then it asks if the user would like to name every player one by one? If user chooses YES,  it further asks what name to give. If the user chooses NO, it simply names the player as default name with the number following it. Lastly all the players' names are displayed as a list, with players' number and name, before actually starting the game.